Bug fixes, refactoring, and obtaining of mood values

+++ addon.py
@@ -8,7 +8,7 @@ from src import json_utils
 addon = xbmcaddon.Addon()
 addonname = addon.getAddonInfo('name')
 dialog = xbmcgui.Dialog()
-window = pyxbmct.AddonDialogWindow('Kodi Episode Picker')
+window = pyxbmct.AddonDialogWindow(addonname)
 
 # User configuration data that will be used throughout the program
 preset_name = ""
@@ -21,19 +21,22 @@ languages_choice = []
 parental_advisory_choice = []
 closed_captions_choice = False
 
-mood_choice = []
+mood_choice = {}
 
 # Displays main page where the user can choose to randomly select, configure, or choose preset
 def display_home_page():
     # user_choice is assigned based on the user's choice. 
+    # -1 : Canceled
     # 0 : Presets
     # 1 : Random Select
     # 2 : Search Across Streaming Services
-    user_choice = dialog.yesnocustom("Kodi Episode Picker", "What would you like to do?", "Search Services", "Presets", "Random Select")
-    if user_choice == 0:
+    user_choice = dialog.yesnocustom(addonname, "What would you like to do?", "Search Services", "Presets", "Random Select")
+    if user_choice == -1:
+        xbmc.executebuiltin("Dialog.Close(%s, true)" % dialog) 
+    elif user_choice == 0:
         display_preset_page()
     elif user_choice == 1:
-        user_choice = dialog.yesno("Kodi Episode Picker", "Would you like to choose media based on your mood today?")
+        user_choice = dialog.yesno(addonname, "Would you like to choose media based on your mood today?")
         if user_choice:
             display_mood_page()
         else:
@@ -50,7 +53,9 @@ def display_preset_page():
     preset_labels.insert(0, "Delete Preset")
     preset_labels.insert(0, "Add Preset")
     preset_choice_num = dialog.select("Choose Preset", preset_labels)
-    if preset_choice_num == 0:
+    if preset_choice_num == -1:
+        display_home_page()
+    elif preset_choice_num == 0:
         display_configuration_page()
     elif preset_choice_num == 1: 
         display_preset_delete_page()
@@ -59,26 +64,34 @@ def display_preset_page():
     else:
         preset_choice = preset_labels[preset_choice_num]
         json_utils.update_current_config(preset_choice)
-        dialog.ok("Kodi Episode Picker", f"Your random play button is now configured to {preset_choice}!")
+        dialog.ok(addonname, f"Your random play button is now configured to {preset_choice}!")
+        display_home_page()
 
 def display_preset_delete_page():
     preset_names = json_utils.get_preset_names()
     preset_labels = [f"{preset}" for preset in preset_names]
     preset_choice_num = [dialog.select("Choose what preset to delete", preset_labels)]
-    preset_choice = convert_indexes_to_strings(preset_choice_num, preset_labels)[0]
-    json_utils.delete_preset(preset_choice)
-    if preset_choice == json_utils.get_current_config()["name"]:
-        json_utils.delete_current_config()
-    dialog.ok("Kodi Episode Picker", f"You have deleted the preset {preset_choice}!")
+    if preset_choice_num[0] != -1:
+        preset_choice = convert_indexes_to_strings(preset_choice_num, preset_labels)[0]
+        json_utils.delete_preset(preset_choice)
+        if preset_choice == json_utils.get_current_config()["name"]:
+            json_utils.delete_current_config()
+        dialog.ok(addonname, f"You have deleted the preset {preset_choice}!")
+        display_preset_page()
+    else:
+        display_preset_page()
 
 def display_preset_edit_page():
     preset_names = json_utils.get_preset_names()
     preset_labels = [f"{preset}" for preset in preset_names]
     preset_choice_num = [dialog.select("Choose what preset to edit", preset_labels)]
-    preset_choice = convert_indexes_to_strings(preset_choice_num, preset_labels)[0]
-    preset_dict = json_utils.get_preset(preset_choice)
-    load_preset(preset_dict)
-    display_configuration_page()
+    if preset_choice_num[0] != -1:
+        preset_choice = convert_indexes_to_strings(preset_choice_num, preset_labels)[0]
+        preset_dict = json_utils.get_preset(preset_choice)
+        load_preset(preset_dict)
+        display_configuration_page()
+    else:
+        display_preset_page()
 
 # Displays the configuration home page       
 def display_configuration_page():
@@ -117,6 +130,8 @@ def display_configuration_page():
 
 def display_streaming_service_page():
     media_name = dialog.input("What media do you want to find today?")
+    if media_name == "":
+        display_home_page()
 
 def display_mood_page():
     window.setGeometry(400, 400, 11, 1)
@@ -143,14 +158,14 @@ def display_mood_page():
     submit_button = pyxbmct.addonwindow.Button('Submit', font='font14')
     window.placeControl(submit_button, 10, 0, rowspan=1, columnspan=1)
     window.connect(submit_button, lambda:
-        play_random_media_with_mood()
+        play_random_media_with_mood(happy_sad_slider, scary_relaxed_slider, romantic_platonic_slider, mindless_thought_provoking_slider, slow_paced_action_packed_slider)
     )
     window.doModal()
 
 # Displays a page where the user can choose if they want to see media they've already seen
 def display_watch_status_page():
     global watch_status_choice
-    watch_status_choice = dialog.yesnocustom("Kodi Episode Picker", "Do you want to see media you've already watched?", "Unwatched", "Any", "Watched")
+    watch_status_choice = dialog.yesnocustom(addonname, "Do you want to see media you've already watched?", "Unwatched", "Any", "Watched")
 
 # Displays a page where the user can choose what genres they want to watch
 def display_genre_page():
@@ -158,7 +173,10 @@ def display_genre_page():
     genres = ["Action", "Adventure" , "Animation" , "Comedy" , "Crime" , "Drama" , "Fantasy" , "Horror", "Mystery" , "Romance" , "Sci-Fi" , "Western" , "Thriller" , "Documentary" , "History" , "Sport"]
     genre_labels = [f"{genre}" for genre in genres]
     genre_choice_nums = dialog.multiselect("Select Movie Genres", genre_labels)
-    genre_choice = convert_indexes_to_strings(genre_choice_nums, genre_labels)
+    if genre_choice_nums == None:
+        pass
+    else:    
+        genre_choice = convert_indexes_to_strings(genre_choice_nums, genre_labels)
 
 # Displays a page where the user can choose what lengths of media they want to watch
 def display_length_page():
@@ -166,7 +184,10 @@ def display_length_page():
     lengths = ["0-30 Minutes", "31-60 minutes", "61-90 minutes","91-120 minutes","121-180 minutes",">180 minutes"]
     length_labels = [f"{length}" for length in lengths]
     length_choice_nums = dialog.multiselect("Select Movie Length", length_labels)
-    length_choice = convert_indexes_to_strings(length_choice_nums, length_labels)
+    if length_choice_nums == None:
+        pass
+    else:
+        length_choice = convert_indexes_to_strings(length_choice_nums, length_labels)
 
 # Displays a page where the user can choose what release years they would like to watch
 def display_release_year_page():
@@ -174,12 +195,15 @@ def display_release_year_page():
     min_year = 1900
     max_year = 2030
     years = [str(year) for year in range(min_year, max_year + 1)][::-1]
-    min_year_input = max_year - dialog.select("Choose Minimum Year", years)
+    min_year_input = min_year - dialog.select("Choose Minimum Year", years)
     max_year_input = max_year - dialog.select("Choose Maximum Year", years)
+    if min_year_input == min_year + 1:
+        min_year_input = min_year
+    if max_year_input == max_year + 1:
+        max_year_input = max_year
     release_year_choice = []
     release_year_choice.append(min_year_input)
     release_year_choice.append(max_year_input)
-
 
 # Displays a page where the user can choose what ratings they would like to watch
 def display_ratings_page():
@@ -187,7 +211,10 @@ def display_ratings_page():
     ratings = ["0-1", "1.1-2", "2.1-3", "3.1-4", "4.1-5", "5.1-6", "6.1-7", "7.1-8", "8.1-9", "9.1-10"]
     ratings_labels = [f"{rating}" for rating in ratings]
     ratings_choice_nums = dialog.multiselect("Select Movie Ratings", ratings_labels)
-    ratings_choice = convert_indexes_to_strings(ratings_choice_nums, ratings_labels)
+    if ratings_choice_nums == None:
+        pass
+    else:
+        ratings_choice = convert_indexes_to_strings(ratings_choice_nums, ratings_labels)
 
 # Displays a page where the user can choose what languages they would like to watch
 def display_languages_page():
@@ -195,7 +222,10 @@ def display_languages_page():
     languages = ["English", "Spanish", "German", "French", "Mandarin", "Hindi", "Arabic", "Portuguese", "Bengali", "Russian", "Japanese"]
     language_labels = [f"{language}" for language in languages]
     language_choice_nums = dialog.multiselect("Select Movie Languages", language_labels)
-    languages_choice = convert_indexes_to_strings(language_choice_nums, language_labels)
+    if language_choice_nums == None:
+        pass
+    else:
+        languages_choice = convert_indexes_to_strings(language_choice_nums, language_labels)
 
 # Displays a page where the user can choose what parental advisory ratings they would like to watch
 def display_parental_advisory_page():
@@ -203,12 +233,15 @@ def display_parental_advisory_page():
     parent_ratings = ["G", "PG", "PG-13", "R", "NR"]
     parent_rating_labels = [f"{parent_rating}" for parent_rating in parent_ratings]
     parent_rating_choice_nums = dialog.multiselect("Select Parental Advisory Ratings", parent_rating_labels)
-    parental_advisory_choice = convert_indexes_to_strings(parent_rating_choice_nums, parent_rating_labels)
+    if parent_rating_choice_nums == None:
+        pass
+    else:
+        parental_advisory_choice = convert_indexes_to_strings(parent_rating_choice_nums, parent_rating_labels)
 
 # Displays a page where the user can choose if they want to only watch media with closed captions
 def display_closed_captions_page():
     global closed_captions_choice
-    closed_captions_choice = dialog.yesno("Kodi Episode Picker", "Do you want to only watch media with closed captions?")
+    closed_captions_choice = dialog.yesno(addonname, "Do you want to only watch media with closed captions?")
 
 # Controls for user to save configuration as preset
 def display_preset_save_page():
@@ -225,22 +258,27 @@ def display_preset_save_page():
         json_utils.replace_preset(preset_dict)
     json_utils.update_current_config(preset_name)
     reset_config()
-    dialog.ok("Kodi Episode Picker","Your preset is now ready to play random media!")
+    dialog.ok(addonname,"Your preset is now ready to play random media!")
 
 def display_preset_exit_page():
-    user_choice = dialog.yesno("Kodi Episode Picker", "Do you want to stop configuring this preset? Your current configurations won't be saved.")
+    user_choice = dialog.yesno(addonname, "Do you want to stop configuring this preset? Your current configurations won't be saved.")
     if user_choice:
-        display_home_page()
+        display_preset_page()
     else:
         display_configuration_page()
 
 # This function will use the user configuration data arrays to randomly choose from media
 def play_random_media():
-    dialog.ok("Kodi Episode Picker","This is where our app will randomly play a movie.")
+    dialog.ok(addonname, "This is where our app will randomly play a movie.")
 
 # This function will use the user configuration data arrays to randomly choose from media
-def play_random_media_with_mood():
-    dialog.ok("Kodi Episode Picker","This is where our app will randomly play a movie.")
+def play_random_media_with_mood(happy_sad, scary_relaxed, romantic_platonic, mindless_thought_provoking, slow_paced_action_packed):
+    dialog.ok(addonname, "Happy/Sad value: " + str(happy_sad.getPercent()) + "\n" +
+                         "Scary/Relaxed value: " + str(scary_relaxed.getPercent()) + "\n" +
+                         "Romantic/Platonic value: " + str(romantic_platonic.getPercent()) + "\n" +
+                         "Mindless/Thought Provoking value: " + str(mindless_thought_provoking.getPercent()) + "\n" +
+                         "Slow Paced/Action Packed value: " + str(slow_paced_action_packed.getPercent()) + "\n")
+    dialog.ok(addonname, "This is where our app will randomly play a movie.")
 
 # This function takes the indices that the multiselect controls produce and fills 
 # them with the corresponding strings. This will make it much easier to program with
